==> notebooks/01_Adding_Continents.ipynb <==
# Adding continents to the countries in the suicide info

First, loading the data from the website, scraping it after and creating a dictionary.

Then load the csv into a df and add the continent column.

import requests
from bs4 import BeautifulSoup
import pandas as pd

url = "https://www.worldatlas.com/cntycont.htm"

atlas = BeautifulSoup(requests.get(url).content, "html.parser")

world = pd.DataFrame()

overview = atlas.find("div", class_="miscTxt")

continents = overview.find_all("h2")

countries = overview.find_all("ul")

continent_list = ([item.text
                  .replace("\n","")
                  .replace("(","")
                  .replace(")","")
                  .replace("1","")
                  .replace("2","")
                  .replace("3","")
                  .replace("4","")
                  .replace("5","")
                  .replace("6","")
                  .replace("7","")
                  .replace("8","")
                  .replace("9","")
                  .replace("0","")
                  .strip()
                  .capitalize()
                  for item in continents]
                )

continent_list = continent_list[:-2]

dictionary = {}
for i in range(len(continent_list)):
    country_list = [item.text for item in countries[i].find_all("li")]
    for items in country_list:
        dictionary[items]=continent_list[i]

dictionary["Aruba"] = "S. america"

dictionary["Cabo Verde"] = "Africa"

dictionary["Macau"] = "Asia"

dictionary["Puerto Rico"] = "N. america"

dictionary["Republic of Korea"] = "Asia"

suicides = pd.read_csv("../data/master.csv")

suicides = suicides.assign(continent = None)

suicides.shape

suicides = suicides[suicides["country"]!='Saint Vincent and Grenadines'].reset_index(drop=True)

for i in range(len(suicides)):
    suicides.loc[i, "continent"]=dictionary[suicides.loc[i,"country"]]

suicides.continent.value_counts()

suicides.to_csv("../data/master.csv", index=False)
