==> notebooks/01b_three_letter_country_dict.ipynb <==
import requests
import pandas as pd
from bs4 import BeautifulSoup

url = "https://laendercode.net/en/3-letter-list.html"

laender = BeautifulSoup(requests.get(url).content, "html.parser")

land_list = laender.find_all("tr")

for i in range(1,len(land_list)):
    land_list[i]= land_list[i].find_all("td")

dictionary = {}

parse = [land_list[i] for i in range(1,len(land_list))]

for items in parse:
    dictionary[items[0].text]=items[1].text

dictionary
