==> notebooks/02_economic_information.ipynb <==
import pandas as pd
import numpy as np

### Household debt

house_debt = pd.read_csv("../data/hausehold_debt.csv")

house_debt.head()

house_debt.isna().sum()

#check how many locations

house_debt["LOCATION"].value_counts()

# check if equal time for all locations

house_debt["TIME"].value_counts()

house_debt["INDICATOR"].value_counts()

house_debt["SUBJECT"].value_counts()

house_debt["MEASURE"].value_counts()

house_debt["FREQUENCY"].value_counts()

# drop columns 	INDICATOR	SUBJECT	MEASURE	FREQUENCY flag codes

house_debt = house_debt.drop(columns = ["Flag Codes", "INDICATOR", "SUBJECT", "MEASURE", "FREQUENCY"])

# rename columns LOCATION	TIME	Value

house_debt = house_debt.rename(columns = {"LOCATION" : "country", "TIME" : "year", "Value" : "household debt"})

### Household spending

house_spending = pd.read_csv("../data/haushold_spending.csv")

house_spending

house_spending.isna().sum()

house_spending["LOCATION"].value_counts()

# house_spending["TIME"].value_counts()

house_spending = house_spending.drop(columns = ["Flag Codes", "INDICATOR", "SUBJECT", "MEASURE", "FREQUENCY"])

house_spending = house_spending.rename(columns = {"LOCATION" : "country", "TIME" : "year", "Value" : "household spent"})

house_spending.head()

### Poverty rate

poverty = pd.read_csv("../data/poverty_rate.csv")

poverty.head()

poverty.isna().sum()

# poverty["LOCATION"].value_counts()

# in this case subject contain the age group (0_17 / 18_65 / 66MORE / TOT: Total)
# decided to keep the age group, but change the values and column name.

poverty["SUBJECT"].value_counts()

poverty["SUBJECT"] = poverty["SUBJECT"].replace({"0_17": "0-17", "18_65" : "18-65", "66MORE" : "66+", "TOT": "total"})

poverty["INDICATOR"].value_counts()

poverty["MEASURE"].value_counts()

poverty["FREQUENCY"].value_counts()

poverty = poverty.drop(columns = ["Flag Codes", "INDICATOR", "MEASURE", "FREQUENCY"])

poverty.head()

poverty = poverty.rename(columns = {"LOCATION" : "country", "TIME" : "year", "Value" : "poverty rate", "SUBJECT": "age group"})

### Internet access


Percentages of households who has internet access

int_access = pd.read_csv("../data/internet_access.csv")

int_access.head()

# int_access["LOCATION"].value_counts()

int_access.isna().sum()

int_access["INDICATOR"].value_counts()

int_access["SUBJECT"].value_counts()

int_access["MEASURE"].value_counts()

int_access["FREQUENCY"].value_counts()

int_access = int_access.drop(columns = ["INDICATOR", "SUBJECT", "MEASURE", "FREQUENCY", "Flag Codes"])

int_access = int_access.rename(columns = {"LOCATION" : "country", "TIME" : "year", "Value" : "internet access"})

### Labour compensation per hour

lab_hour = pd.read_csv("../data/labour_compensation_per_hour.csv")

lab_hour.head()

# lab_hour["LOCATION"].value_counts()

lab_hour["INDICATOR"].value_counts()


lab_hour["SUBJECT"].value_counts()

lab_hour["MEASURE"].value_counts()

#dropping rows for anual growth and keeping rows for 2010 = 100

lab_hour = lab_hour.drop(lab_hour[(lab_hour.MEASURE == "IDX2010")].index)

lab_hour["MEASURE"].value_counts()

lab_hour["FREQUENCY"].value_counts()

lab_hour = lab_hour.drop(columns = ["INDICATOR", "SUBJECT", "MEASURE", "FREQUENCY", "Flag Codes"])

lab_hour = lab_hour.rename(columns = {"LOCATION" : "country", "TIME" : "year", "Value" : "labour comp per hour"})

### Unenployment Rate

unemp_rate = pd.read_csv("../data/unemployment_rate.csv")

unemp_rate.head()

unemp_rate.isna().sum()

# unemp_rate["LOCATION"].value_counts()

unemp_rate["INDICATOR"].value_counts()

unemp_rate["SUBJECT"].value_counts()

unemp_rate["MEASURE"].value_counts()

unemp_rate["FREQUENCY"].value_counts()

# droping extra rows for frequency and keeping just anual values

unemp_rate = unemp_rate.drop(unemp_rate[(unemp_rate.FREQUENCY == "M")].index)

unemp_rate = unemp_rate.drop(unemp_rate[unemp_rate.FREQUENCY == "Q"].index)
unemp_rate["FREQUENCY"].value_counts()

#dropping extra rows for subject and keeping just total

unemp_rate = unemp_rate.drop(unemp_rate[(unemp_rate.SUBJECT == "MEN")].index)

unemp_rate = unemp_rate.drop(unemp_rate[unemp_rate.SUBJECT == "WOMEN"].index)

unemp_rate["SUBJECT"].value_counts()

unemp_rate = unemp_rate.drop(columns = ["INDICATOR", "SUBJECT", "MEASURE", "FREQUENCY", "Flag Codes"])


unemp_rate = unemp_rate.rename(columns = {"LOCATION" : "country", "TIME" : "year", "Value" : "unemployment rate"})

unemp_rate.reset_index(drop = True)

unemp_rate.dtypes

unemp_rate["year"] = unemp_rate["year"].astype(int)

unemp_rate.dtypes

house_debt = house_debt.merge(lab_hour, how = 'left',  on = ["country", "year"]).reset_index(drop = True)

house_debt = house_debt.merge(poverty, how = 'left', on= ["country", "year"]).reset_index(drop = True)

house_debt.isna().sum()

house_debt = house_debt.merge(int_access, how = 'left', on= ["country", "year"]).reset_index(drop = True)

house_debt

economic_info = house_debt.merge(unemp_rate, how = 'left', on= ["country", "year"]).reset_index(drop = True)

economic_info.isna().sum()

len(economic_info)

country_code = {'AFG': 'Afghanistan','ALA': 'Aland Islands','ALB': 'Albania','DZA': 'Algeria','ASM': 'American Samoa','AND': 'Andorra','AGO': 'Angola','AIA': 'Anguilla','ATA': 'Antarctica','ATG': 'Antigua and Barbuda','ARG': 'Argentina','ARM': 'Armenia','ABW': 'Aruba','AUS': 'Australia','AUT': 'Austria','AZE': 'Azerbaijan','BHS': 'Bahamas','BHR': 'Bahrain','BGD': 'Bangladesh','BRB': 'Barbados','BLR': 'Belarus','BEL': 'Belgium','BLZ': 'Belize','BEN': 'Benin','BMU': 'Bermuda','BTN': 'Bhutan','BOL': 'Bolivia','BES': 'Bonaire, Saint Eustatius and Saba ','BIH': 'Bosnia and Herzegovina','BWA': 'Botswana','BVT': 'Bouvet Island','BRA': 'Brazil','IOT': 'British Indian Ocean Territory','VGB': 'British Virgin Islands','BRN': 'Brunei','BGR': 'Bulgaria','BFA': 'Burkina Faso','BDI': 'Burundi','KHM': 'Cambodia','CMR': 'Cameroon','CAN': 'Canada','CPV': 'Cape Verde','CYM': 'Cayman Islands','CAF': 'Central African Republic','TCD': 'Chad','CHL': 'Chile','CHN': 'China','CXR': 'Christmas Island','CCK': 'Cocos Islands','COL': 'Colombia','COM': 'Comoros','COK': 'Cook Islands','CRI': 'Costa Rica','HRV': 'Croatia','CUB': 'Cuba','CUW': 'Curacao','CYP': 'Cyprus','CZE': 'Czech Republic','COD': 'Democratic Republic of the Congo','DNK': 'Denmark','DJI': 'Djibouti','DMA': 'Dominica','DOM': 'Dominican Republic','TLS': 'East Timor','ECU': 'Ecuador','EGY': 'Egypt','SLV': 'El Salvador','GNQ': 'Equatorial Guinea','ERI': 'Eritrea','EST': 'Estonia','ETH': 'Ethiopia','FLK': 'Falkland Islands','FRO': 'Faroe Islands','FJI': 'Fiji','FIN': 'Finland','FRA': 'France','GUF': 'French Guiana','PYF': 'French Polynesia','ATF': 'French Southern Territories','GAB': 'Gabon','GMB': 'Gambia','GEO': 'Georgia','DEU': 'Germany','GHA': 'Ghana','GIB': 'Gibraltar','GRC': 'Greece','GRL': 'Greenland','GRD': 'Grenada','GLP': 'Guadeloupe','GUM': 'Guam','GTM': 'Guatemala','GGY': 'Guernsey','GIN': 'Guinea','GNB': 'Guinea-Bissau','GUY': 'Guyana','HTI': 'Haiti','HMD': 'Heard Island and McDonald Islands','HND': 'Honduras','HKG': 'Hong Kong','HUN': 'Hungary','ISL': 'Iceland','IND': 'India','IDN': 'Indonesia','IRN': 'Iran','IRQ': 'Iraq','IRL': 'Ireland','IMN': 'Isle of Man','ISR': 'Israel','ITA': 'Italy','CIV': 'Ivory Coast','JAM': 'Jamaica','JPN': 'Japan','JEY': 'Jersey','JOR': 'Jordan','KAZ': 'Kazakhstan','KEN': 'Kenya','KIR': 'Kiribati','XKX': 'Kosovo','KWT': 'Kuwait','KGZ': 'Kyrgyzstan','LAO': 'Laos','LVA': 'Latvia','LBN': 'Lebanon','LSO': 'Lesotho','LBR': 'Liberia','LBY': 'Libya','LIE': 'Liechtenstein','LTU': 'Lithuania','LUX': 'Luxembourg','MAC': 'Macao','MKD': 'Macedonia','MDG': 'Madagascar','MWI': 'Malawi','MYS': 'Malaysia','MDV': 'Maldives','MLI': 'Mali','MLT': 'Malta','MHL': 'Marshall Islands','MTQ': 'Martinique','MRT': 'Mauritania','MUS': 'Mauritius','MYT': 'Mayotte','MEX': 'Mexico','FSM': 'Micronesia','MDA': 'Moldova','MCO': 'Monaco','MNG': 'Mongolia','MNE': 'Montenegro','MSR': 'Montserrat','MAR': 'Morocco','MOZ': 'Mozambique','MMR': 'Myanmar','NAM': 'Namibia','NRU': 'Nauru','NPL': 'Nepal','NLD': 'Netherlands','ANT': 'Netherlands Antilles','NCL': 'New Caledonia','NZL': 'New Zealand','NIC': 'Nicaragua','NER': 'Niger','NGA': 'Nigeria','NIU': 'Niue','NFK': 'Norfolk Island','PRK': 'North Korea','MNP': 'Northern Mariana Islands','NOR': 'Norway','OMN': 'Oman','PAK': 'Pakistan','PLW': 'Palau','PSE': 'Palestinian Territory','PAN': 'Panama','PNG': 'Papua New Guinea','PRY': 'Paraguay','PER': 'Peru','PHL': 'Philippines','PCN': 'Pitcairn','POL': 'Poland','PRT': 'Portugal','PRI': 'Puerto Rico','QAT': 'Qatar','COG': 'Republic of the Congo','REU': 'Reunion','ROU': 'Romania','RUS': 'Russia','RWA': 'Rwanda','BLM': 'Saint Barthelemy','SHN': 'Saint Helena','KNA': 'Saint Kitts and Nevis','LCA': 'Saint Lucia','MAF': 'Saint Martin','SPM': 'Saint Pierre and Miquelon','VCT': 'Saint Vincent and the Grenadines','WSM': 'Samoa','SMR': 'San Marino','STP': 'Sao Tome and Principe','SAU': 'Saudi Arabia','SEN': 'Senegal','SRB': 'Serbia','SCG': 'Serbia and Montenegro','SYC': 'Seychelles','SLE': 'Sierra Leone','SGP': 'Singapore','SXM': 'Sint Maarten','SVK': 'Slovakia','SVN': 'Slovenia','SLB': 'Solomon Islands','SOM': 'Somalia','ZAF': 'South Africa','SGS': 'South Georgia and the South Sandwich Islands','KOR': 'REPUBLIC OF KOREA','SSD': 'South Sudan','ESP': 'Spain','LKA': 'Sri Lanka','SDN': 'Sudan','SUR': 'Suriname','SJM': 'Svalbard and Jan Mayen','SWZ': 'Swaziland','SWE': 'Sweden','CHE': 'Switzerland','SYR': 'Syria','TWN': 'Taiwan','TJK': 'Tajikistan','TZA': 'Tanzania','THA': 'Thailand','TGO': 'Togo','TKL': 'Tokelau','TON': 'Tonga','TTO': 'Trinidad and Tobago','TUN': 'Tunisia','TUR': 'Turkey','TKM': 'Turkmenistan','TCA': 'Turks and Caicos Islands','TUV': 'Tuvalu','VIR': 'U.S. Virgin Islands','UGA': 'Uganda','UKR': 'Ukraine','ARE': 'United Arab Emirates','GBR': 'United Kingdom','USA': 'United States','UMI': 'United States Minor Outlying Islands','URY': 'Uruguay','UZB': 'Uzbekistan','VUT': 'Vanuatu','VAT': 'Vatican','VEN': 'Venezuela','VNM': 'Vietnam','WLF': 'Wallis and Futuna','ESH': 'Western Sahara','YEM': 'Yemen','ZMB': 'Zambia','ZWE': 'Zimbabwe'}

#for country in economic_info["country"]:
#    if country in country_code:
#        economic_info["country"] = country

for i in range(len(economic_info)):
    economic_info.loc[i, "country"] = country_code[economic_info.loc[i, "country"]]

economic_info

# economic_info.loc[economic_info["country"] == "REPUBLIC OF KOREA"]

economic_info.to_csv("../data/economic_info.csv", index = False)

economic_info = pd.read_csv("../data/economic_info.csv")

economic_info.head()
