==> notebooks/03_alcohol-consumption-cleaning.ipynb <==
import pandas as pd
import numpy as np


### Korea


korea_alc = pd.read_excel("data/alcohol-korea.xlsx")

korea_alc.head()

korea_alc["Unnamed: 1"].isna().sum()

# check if rows with nan values are important

korea_na = korea_alc.loc[korea_alc["Unnamed: 1"].isna()]

korea_na

# dropping nan and reseting index

korea_alc = korea_alc.dropna(inplace = False).reset_index(drop = True)

# rename the columns

korea_alc = korea_alc.rename(columns = {"Spirits Consumption Per Capita in South Korea" : "Year", "Unnamed: 1" : "Alcohol per Capita"})

korea_alc = korea_alc.drop(korea_alc.index[0]).reset_index(drop = True)

korea_alc.head()

# sorting values ascending by year

korea_alc = korea_alc.sort_values(by = "Year").reset_index(drop = True)

# Create new column with country name

korea_alc["country"] = "REPUBLIC OF KOREA"

korea_alc.head()

# save dataframe

korea_alc.to_csv("data/korea_alcohol_cons.csv", index = False)

### US

usa_alc = pd.read_excel("data/alcohol-US.xlsx")

usa_alc

usa_alc["Unnamed: 1"].isna().sum()

# check if rows with nan values are important

usa_na = usa_alc.loc[usa_alc["Unnamed: 1"].isna()]
usa_na

# dropping nan and reseting index

usa_alc = usa_alc.dropna(inplace = False).reset_index(drop = True)

usa_alc.head()

# rename the columns

usa_alc = usa_alc.rename(columns = {"Spirits Consumption Per Capita in the US" : "Year", "Unnamed: 1" : "Alcohol per Capita"})

usa_alc.head()

# dropping first row

usa_alc = usa_alc.drop(usa_alc.index[0]).reset_index(drop = True)

usa_alc.head()

# sorting values ascending by year

usa_alc = usa_alc.sort_values(by = "Year").reset_index(drop = True)

usa_alc.head()

# Create new column with country name
usa_alc["country"] = "UNITED STATES"

usa_alc.head()

usa_alc.to_csv("data/usa_alcohol_cons.csv", index = False)

### Russia

russia_alc = pd.read_excel("data/alcohol-Russia.xlsx")

russia_alc.head()

# check nan

russia_alc["Unnamed: 1"].isna().sum()

# check if rows with nan values are important
russia_na = russia_alc.loc[russia_alc["Unnamed: 1"].isna()]

russia_na

# dropping nan and reseting index

russia_alc = russia_alc.dropna(inplace = False).reset_index(drop = True)

russia_alc.head()

# rename the columns
russia_alc = russia_alc.rename(columns = {"Spirits Consumption Per Capita in Russia" : "year", "Unnamed: 1" : "alcohol per capita"})


# dropping first row
russia_alc = russia_alc.drop(russia_alc.index[0]).reset_index(drop = True)

# sorting values ascending by year
russia_alc = russia_alc.sort_values(by = "year").reset_index(drop = True)

russia_alc.head()

# Create new column with country name

russia_alc["country"] = "RUSSIAN FEDERATION"

russia_alc.head()

russia_alc.to_csv("data/russia_alcohol_cons.csv", index = False)

### Japan

japan_alc = pd.read_excel("data/alcohol_cons_japan.xlsx")

japan_alc.head()

# check nan

japan_alc["Unnamed: 1"].isna().sum()

# check if rows with nan values are important
japan_na = japan_alc.loc[japan_alc["Unnamed: 1"].isna()]

japan_na

# dropping nan and reseting index

japan_alc = japan_alc.dropna(inplace = False).reset_index(drop = True)

japan_alc.head()

# rename the columns
japan_alc = japan_alc.rename(columns = {"Spirits Consumption Per Capita in Japan" : "year", "Unnamed: 1" : "alcohol per capita"})


# dropping first row
japan_alc = japan_alc.drop(japan_alc.index[0]).reset_index(drop = True)

# sorting values ascending by year
japan_alc = japan_alc.sort_values(by = "year").reset_index(drop = True)

japan_alc.head()

# Create new column with country name

japan_alc["country"] = "JAPAN"

japan_alc.head()

japan_alc.to_csv("data/japan_alcohol_cons.csv", index = False)
