==> notebooks/03b_alcohol_per_capita_2.ipynb <==
import pandas as pd
import numpy as np

korea = pd.read_csv("data/korea_alcohol_cons.csv")
japan = pd.read_csv("data/japan_alcohol_cons.csv")
russia = pd.read_csv("data/russia_alcohol_cons.csv")
usa = pd.read_csv("data/usa_alcohol_cons.csv")

korea.head()

korea = korea.rename(columns = {"Year" : "year", "Alcohol per Capita" : "alcohol per capita"})
korea.head()

japan.head()

russia.head()

usa.head()

usa = usa.rename(columns = {"Year" : "year", "Alcohol per Capita" : "alcohol per capita"})

usa.head()

countries_all = pd.concat([korea, usa, russia, japan]).reset_index(drop = True)

countries_all

countries_all.to_csv("data/alcohol_per_capita2.csv", index = False)
