==> notebooks/04_alcohol-consumption-merged.ipynb <==
import pandas as pd
import numpy as np

korea_alc = pd.read_csv("data/korea_alcohol_cons.csv")
usa_alc = pd.read_csv("data/usa_alcohol_cons.csv")
russia_alc = pd.read_csv("data/russia_alcohol_cons.csv")

countries_all = pd.concat([korea_alc, usa_alc, russia_alc]).reset_index(drop = True)

countries_all

countries_all.drop(columns = "Unnamed: 0")

countries_all.to_csv("data/alcohol_per_capita.csv")
