==> notebooks/07_health_expenditure_suicide_correlation.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import statsmodels.api as sm

#We found a data set with information on the expenditure on health per country
#as a percentage of the GDP

df_health=pd.read_csv("../data/expenditure_on_health_total_of_gdp.csv")

df_health.head()

df_h = df_health.drop(columns=["id",
                               "utc_created",
                               "utc_updated",
                               "country_code",
                               "indicator_code",
                               "indicator"])

df_h = df_h.rename(columns={"value":"%GDP_health"})
df_h.head()
df_h[df_h["country"] == "Albania"]

#We want to merge this new table with our master table
# Is there a correlation between expenditure on health and suicides rate?

df_suicides = pd.read_csv("../data/master_grouped_ccy.csv")
df_s = df_suicides.drop(columns="Unnamed: 0")
df_s.head()

df_sh = df_s.merge(df_h, how="left", on=["country","year"])

df_sh.head()

#Now we have our tables together, so we can do the linear regression

df_sh = df_sh[df_sh["%GDP_health"].notna()]
df_sh.head()

df_c = df_sh[df_sh["continent"] == "Europe"]
df_sh_cor = sm.add_constant(df_c)

df_sh_cor.head()

y = df_sh_cor["s_rate"]
X = df_sh_cor[["const","%GDP_health"]]
lin_reg = sm.OLS(y,X)
results = lin_reg.fit()
results.summary()
