==> notebooks/08_subsetting_oecd_countries.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("../data/master_grouped_plusage.csv")
df = df.drop(columns="Unnamed: 0")
df.head()

oecd = ["AUSTRALIA",
"AUSTRIA",
"BELGIUM",
"CANADA",
"CHILE",
"COLOMBIA",
"CZECH REPUBLIC",
"DENMARK",
"ESTONIA",
"FINLAND",
"FRANCE",
"GERMANY",
"GREECE",
"HUNGARY",
"ICELAND",
"IRELAND",
"ISRAEL",
"ITALY",
"JAPAN",
"REPUBLIC OF KOREA",
"LATVIA",
"LITHUANIA",
"LUXEMBOURG",
"MEXICO",
"NETHERLANDS",
"NEW ZEALAND",
"NORWAY",
"POLAND",
"PORTUGAL",
"SLOVAKIA",
"SLOVENIA",
"SPAIN",
"SWEDEN",
"SWITZERLAND",
"TURKEY",
"UNITED KINGDOM",
"UNITED STATES"]

len(oecd)

df["country"] = df["country"].str.upper()

df["oecd"] = np.where(df["country"].isin(oecd), 1, 0)

df[df["country"]== "SLOVAKIA"]

df_oecd = df[df["oecd"]==1]

df_oecd.head()

df_oecd.to_csv("../data/master_oecd.csv", index=False)
