==> notebooks/10_OECD_pension_inequality.ipynb <==
import pandas as pd

p = pd.read_csv("../data/pension_spending.csv")
ine = pd.read_csv("../data/inequality.csv")

p.head()

pension = p.drop(columns=["INDICATOR","SUBJECT","MEASURE","FREQUENCY","Flag Codes"])

pension.head()

pension = pension.rename(columns={"LOCATION":"country","TIME":"year","Value":"%gdp_pen"})

pension["country"].value_counts()

country_code = {'OECD':'OECD','AFG': 'Afghanistan','ALA': 'Aland Islands','ALB': 'Albania','DZA': 'Algeria','ASM': 'American Samoa','AND': 'Andorra','AGO': 'Angola','AIA': 'Anguilla','ATA': 'Antarctica','ATG': 'Antigua and Barbuda','ARG': 'Argentina','ARM': 'Armenia','ABW': 'Aruba','AUS': 'Australia','AUT': 'Austria','AZE': 'Azerbaijan','BHS': 'Bahamas','BHR': 'Bahrain','BGD': 'Bangladesh','BRB': 'Barbados','BLR': 'Belarus','BEL': 'Belgium','BLZ': 'Belize','BEN': 'Benin','BMU': 'Bermuda','BTN': 'Bhutan','BOL': 'Bolivia','BES': 'Bonaire, Saint Eustatius and Saba ','BIH': 'Bosnia and Herzegovina','BWA': 'Botswana','BVT': 'Bouvet Island','BRA': 'Brazil','IOT': 'British Indian Ocean Territory','VGB': 'British Virgin Islands','BRN': 'Brunei','BGR': 'Bulgaria','BFA': 'Burkina Faso','BDI': 'Burundi','KHM': 'Cambodia','CMR': 'Cameroon','CAN': 'Canada','CPV': 'Cape Verde','CYM': 'Cayman Islands','CAF': 'Central African Republic','TCD': 'Chad','CHL': 'Chile','CHN': 'China','CXR': 'Christmas Island','CCK': 'Cocos Islands','COL': 'Colombia','COM': 'Comoros','COK': 'Cook Islands','CRI': 'Costa Rica','HRV': 'Croatia','CUB': 'Cuba','CUW': 'Curacao','CYP': 'Cyprus','CZE': 'Czech Republic','COD': 'Democratic Republic of the Congo','DNK': 'Denmark','DJI': 'Djibouti','DMA': 'Dominica','DOM': 'Dominican Republic','TLS': 'East Timor','ECU': 'Ecuador','EGY': 'Egypt','SLV': 'El Salvador','GNQ': 'Equatorial Guinea','ERI': 'Eritrea','EST': 'Estonia','ETH': 'Ethiopia','FLK': 'Falkland Islands','FRO': 'Faroe Islands','FJI': 'Fiji','FIN': 'Finland','FRA': 'France','GUF': 'French Guiana','PYF': 'French Polynesia','ATF': 'French Southern Territories','GAB': 'Gabon','GMB': 'Gambia','GEO': 'Georgia','DEU': 'Germany','GHA': 'Ghana','GIB': 'Gibraltar','GRC': 'Greece','GRL': 'Greenland','GRD': 'Grenada','GLP': 'Guadeloupe','GUM': 'Guam','GTM': 'Guatemala','GGY': 'Guernsey','GIN': 'Guinea','GNB': 'Guinea-Bissau','GUY': 'Guyana','HTI': 'Haiti','HMD': 'Heard Island and McDonald Islands','HND': 'Honduras','HKG': 'Hong Kong','HUN': 'Hungary','ISL': 'Iceland','IND': 'India','IDN': 'Indonesia','IRN': 'Iran','IRQ': 'Iraq','IRL': 'Ireland','IMN': 'Isle of Man','ISR': 'Israel','ITA': 'Italy','CIV': 'Ivory Coast','JAM': 'Jamaica','JPN': 'Japan','JEY': 'Jersey','JOR': 'Jordan','KAZ': 'Kazakhstan','KEN': 'Kenya','KIR': 'Kiribati','XKX': 'Kosovo','KWT': 'Kuwait','KGZ': 'Kyrgyzstan','LAO': 'Laos','LVA': 'Latvia','LBN': 'Lebanon','LSO': 'Lesotho','LBR': 'Liberia','LBY': 'Libya','LIE': 'Liechtenstein','LTU': 'Lithuania','LUX': 'Luxembourg','MAC': 'Macao','MKD': 'Macedonia','MDG': 'Madagascar','MWI': 'Malawi','MYS': 'Malaysia','MDV': 'Maldives','MLI': 'Mali','MLT': 'Malta','MHL': 'Marshall Islands','MTQ': 'Martinique','MRT': 'Mauritania','MUS': 'Mauritius','MYT': 'Mayotte','MEX': 'Mexico','FSM': 'Micronesia','MDA': 'Moldova','MCO': 'Monaco','MNG': 'Mongolia','MNE': 'Montenegro','MSR': 'Montserrat','MAR': 'Morocco','MOZ': 'Mozambique','MMR': 'Myanmar','NAM': 'Namibia','NRU': 'Nauru','NPL': 'Nepal','NLD': 'Netherlands','ANT': 'Netherlands Antilles','NCL': 'New Caledonia','NZL': 'New Zealand','NIC': 'Nicaragua','NER': 'Niger','NGA': 'Nigeria','NIU': 'Niue','NFK': 'Norfolk Island','PRK': 'North Korea','MNP': 'Northern Mariana Islands','NOR': 'Norway','OMN': 'Oman','PAK': 'Pakistan','PLW': 'Palau','PSE': 'Palestinian Territory','PAN': 'Panama','PNG': 'Papua New Guinea','PRY': 'Paraguay','PER': 'Peru','PHL': 'Philippines','PCN': 'Pitcairn','POL': 'Poland','PRT': 'Portugal','PRI': 'Puerto Rico','QAT': 'Qatar','COG': 'Republic of the Congo','REU': 'Reunion','ROU': 'Romania','RUS': 'Russia','RWA': 'Rwanda','BLM': 'Saint Barthelemy','SHN': 'Saint Helena','KNA': 'Saint Kitts and Nevis','LCA': 'Saint Lucia','MAF': 'Saint Martin','SPM': 'Saint Pierre and Miquelon','VCT': 'Saint Vincent and the Grenadines','WSM': 'Samoa','SMR': 'San Marino','STP': 'Sao Tome and Principe','SAU': 'Saudi Arabia','SEN': 'Senegal','SRB': 'Serbia','SCG': 'Serbia and Montenegro','SYC': 'Seychelles','SLE': 'Sierra Leone','SGP': 'Singapore','SXM': 'Sint Maarten','SVK': 'Slovakia','SVN': 'Slovenia','SLB': 'Solomon Islands','SOM': 'Somalia','ZAF': 'South Africa','SGS': 'South Georgia and the South Sandwich Islands','KOR': 'South Korea','SSD': 'South Sudan','ESP': 'Spain','LKA': 'Sri Lanka','SDN': 'Sudan','SUR': 'Suriname','SJM': 'Svalbard and Jan Mayen','SWZ': 'Swaziland','SWE': 'Sweden','CHE': 'Switzerland','SYR': 'Syria','TWN': 'Taiwan','TJK': 'Tajikistan','TZA': 'Tanzania','THA': 'Thailand','TGO': 'Togo','TKL': 'Tokelau','TON': 'Tonga','TTO': 'Trinidad and Tobago','TUN': 'Tunisia','TUR': 'Turkey','TKM': 'Turkmenistan','TCA': 'Turks and Caicos Islands','TUV': 'Tuvalu','VIR': 'U.S. Virgin Islands','UGA': 'Uganda','UKR': 'Ukraine','ARE': 'United Arab Emirates','GBR': 'United Kingdom','USA': 'United States','UMI': 'United States Minor Outlying Islands','URY': 'Uruguay','UZB': 'Uzbekistan','VUT': 'Vanuatu','VAT': 'Vatican','VEN': 'Venezuela','VNM': 'Vietnam','WLF': 'Wallis and Futuna','ESH': 'Western Sahara','YEM': 'Yemen','ZMB': 'Zambia','ZWE': 'Zimbabwe'}

for i in range(len(pension)):
    pension.loc[i, "country"] = country_code[pension.loc[i, "country"]]

pension["country"]=pension["country"].str.upper()

pension.to_csv("../data/pension_df.csv", index=False)

inequality = ine.drop(columns=["INDICATOR","SUBJECT","MEASURE","FREQUENCY","Flag Codes"])

inequality = inequality.rename(columns={"LOCATION":"country","TIME":"year","Value":"inqlt_rate"})

inequality.head()

for i in range(len(inequality)):
    inequality.loc[i, "country"] = country_code[inequality.loc[i, "country"]].upper()

inequality.to_csv("../data/inequality_df.csv", index=False)
