==> notebooks/12b_workforce_workaround.ipynb <==
import pandas as pd
import numpy as np

file = pd.read_csv("../data/working.csv")

file.head()

work = pd.DataFrame({"country":[], "year":[],"work_hours":[]})

work

country_list = file.columns.to_list()

country_list = country_list[1:]

country_list[0]="REPUBLIC OF KOREA"

file.iloc[0,1]

country_df = []
year_df = []
work_hours_df = []
for i in range(len(country_list)): 
    for j in range(len(file)):
        country_df.append(country_list[i])
        year_df.append(file.iloc[j,0])
        work_hours_df.append(file.iloc[j,i+1])
        

work = pd.DataFrame({"country":country_df, "year":year_df,"work_hours":work_hours_df})

work

work.to_csv("../data/work_df.csv",index=False)
