==> notebooks/13_merging_final_dataframe.ipynb <==
import pandas as pd

suicides = pd.read_csv("../data/master_oecd.csv")
alcohol = pd.read_csv("../data/alcohol_per_capita2.csv")
poverty = pd.read_csv("../data/poverty.csv")
working_hours = pd.read_csv("../data/work_df.csv")
economic = pd.read_csv("../data/economic_info.csv")

suicides.head()

alcohol.head()

df1 = suicides.merge(alcohol, how="left", on=["country","year"])

poverty.head()

df2 = df1.merge(poverty, how="left", on=["country","year"])

df2.head()

economic.head()

eco = economic.drop(columns=["age group","poverty rate","internet access"])

eco["country"]=eco["country"].str.upper()

df3 = df2.merge(eco, how="left", on=["country","year"])

df3.head()

working_hours.head()

working_hours["country"]=working_hours["country"].str.upper()

df4 = df3.merge(working_hours, how="left", on=["country","year"])

df4[df4["year"]==2015]

pension = pd.read_csv("../data/pension_df.csv")
inequality= pd.read_csv("../data/inequality_df.csv")

pension.head()

inequality

df5 = df4.merge(pension, how="left", on=["country","year"])

df6 = df5.merge(inequality, how="left", on=["country","year"])

df6[(df6["country"] == "REPUBLIC OF KOREA")&(df6["year"]==2015)&(df6["age"]=="75+ years")]

pov_list = [44.7,44.7,44.7,44.7]

for i in range(len(pov_list)):
    n = 8729
    df6.iloc[n+i,10]=pov_list[i]/100

df6[(df6["country"] == "REPUBLIC OF KOREA")&(df6["year"]==2006)].iloc[0,10]

df6.to_csv("../data/final_df.csv", index=False)

df6["pov_rate"]=df6["pov_rate"]*1000
